requirements.txt
playwright==1.45.0
httpx==0.27.0
tenacity==8.3.0
pydantic==2.8.2
pydantic-settings==2.3.4
gspread==6.0.2
google-auth==2.33.0
google-auth-httplib2==0.2.0
openai==1.35.10
python-dateutil==2.9.0.post0
selectolax==0.3.17
pytest==8.3.2
